project/aidlc-docs/construction/unit4_card_search/src/requirements.txt
# Unit4: Card Search & Display Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1

# Caching
redis==5.0.1
aioredis==2.0.1

# Authentication (Unit1 dependency)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Validation & Serialization
email-validator==2.1.0
python-multipart==0.0.6

# Logging & Monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Unit3 Card Entity Integration
# Note: Will import Card entity from unit3 implementation
